+++ lib/more_libs/setting_v2.0.0/app_setting_screen.dart
@@ -1,7 +1,6 @@
 import 'dart:io';
 
 import 'package:flutter/material.dart';
-import 'package:novel_v3/more_libs/setting_v2.0.0/thancoder_about_widget.dart';
 import 'package:t_widgets/t_widgets.dart';
 import 'package:than_pkg/than_pkg.dart';
 
@@ -62,7 +61,7 @@ class _AppSettingScreenState extends State<AppSettingScreen> {
               // custom image
               _getCustomNovelCoverWidet(),
               //changelog && readme
-              ThancoderAboutWidget(),
+              // ThancoderAboutWidget(),
             ],
           ),
         ),